Algumas mudanças e estou com preguiça de pontuar todas, porem as duas mais importantes são o UPDATE Generico e o Ajuste para pegar um registro no banco via Chave Primaria, antes utilizando FIrstOrDefault ou Where, o desempenho local para buscar o registro era de 700 ms ou mais, ja agora usando FindAsync tenho tempo de 17 ms

+++ Domain/Commands/Base/CommandResult.cs
@@ -14,7 +14,7 @@ namespace Domain.Commands.Base
             Sucesso = sucesso;
         }
 
-        public CommandResult(bool sucesso, int id)
+        public CommandResult(bool sucesso, long id)
         {
             Sucesso = sucesso;
             Id = id;
@@ -38,7 +38,7 @@ namespace Domain.Commands.Base
 
         public bool Sucesso { get; set; }
 
-        public int Id { get; set; }
+        public long Id { get; set; }
 
         public IList<CommandResultErro>? Erros { get; private set; }
     }
